Removed the PowerDetails activity in favor of a MoreDetails page with more information
MoreDetails includes information on power data (current/voltage) and water level data

Redesigned GUI for main activity to incorporate the more details section and reduce the size of power consumption

Fixed the power calculation process and added a method that calculates the space remaining between the two tanks

TODO: connect the buttons over ThingSpeak
      figure out how to format a string to handle up to 2 decimal places (possibly with DecimalFormat)

+++ app/src/main/java/com/example/naffaa/tankcontrol/MoreDetails.java
@@ -59,7 +59,7 @@ public class MoreDetails extends AppCompatActivity {
                             JSONObject inner = outer.getJSONObject(0);
 
                             String[] mDataSet = new String[SIZE];
-                            int HEIGHT = 100; //Adjust height based on size of tank
+                            double HEIGHT = 100; //Adjust height based on size of tank
 
                             mDataSet[0] = inner.getString("field4") + " "; // Tank 1 Water Level
                             mDataSet[1] = inner.getString("field5") + " "; // Tank 2 Water Level
